Refactor useDashboardDropdownOptions
